♻️ refactor(websearch): organize utility imports
- move headers and filter functions to separate modules
- update import paths for better code organization
- maintain consistent header and timeout usage across search implementations
